Slider: update test, rename element and fix style
