+++ src/components/Slider/Container.js
@@ -2,6 +2,6 @@ import styled from 'react-emotion';
 
 export default styled.div`
   position: relative;
-  margin: auto;
   text-align: center;
+  margin: auto auto 20px;
 `;
